feat: add support to delete any pods that might be remaining

+++ controllers/messenger/queues.go
@@ -295,6 +295,10 @@ func (h *Messaging) Consumer(targetName string) { //error {
 						message.Ack(false) // ack to remove from queue
 						return
 					}
+					if del := h.DeletePods(ctx, opLog.WithName("DeletePods"), ns, project, branch); del == false {
+						message.Ack(false) // ack to remove from queue
+						return
+					}
 					if del := h.DeletePVCs(ctx, opLog.WithName("DeletePVCs"), ns, project, branch); del == false {
 						message.Ack(false) // ack to remove from queue
 						return
